+++ memory.h
@@ -4,16 +4,18 @@
 #include <stdio.h>
 #include <stdint.h>
 
+#include "base.h"
+
 typedef struct Arena Arena;
 struct Arena {
 	unsigned char	*buf;
-	size_t		len;
-	size_t		curr_offset;
+	u32		len;
+	u32		curr_offset;
 };
 
-uintptr_t align_forward(uintptr_t ptr, size_t align);
-void *arena_alloc(Arena *a, size_t size);
-void arena_init(Arena *a, void *backing_buffer, size_t backing_buffer_len);
+uintptr_t align_forward(uintptr_t ptr, u32 align);
+void *arena_alloc(Arena *a, u32 size);
+void arena_init(Arena *a, void *backing_buffer, u32 backing_buffer_len);
 void arena_free(Arena *a);
 
 #endif
